issue-55: move lsm core to separate module
