+++ storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/jobs/service/MergeSsTableJob.kt
@@ -5,15 +5,15 @@ import org.quartz.DisallowConcurrentExecution
 import org.quartz.JobExecutionContext
 import org.springframework.scheduling.quartz.QuartzJobBean
 import org.springframework.stereotype.Service
+import ru.itmo.storage.storage.lsm.core.AVLTree
 import ru.itmo.storage.storage.jobs.properties.MergeSstableJobProperties
-import ru.itmo.storage.storage.lsm.MemtableService
-import ru.itmo.storage.storage.lsm.avl.AVLTree
+import ru.itmo.storage.storage.lsm.core.MemtableService
 import ru.itmo.storage.storage.lsm.avl.DefaultAVLTree
-import ru.itmo.storage.storage.lsm.bloomfilter.BloomFilter
-import ru.itmo.storage.storage.lsm.properties.BloomFilterProperties
+import ru.itmo.storage.storage.lsm.core.bloomfilter.BloomFilter
+import ru.itmo.storage.storage.lsm.core.bloomfilter.BloomFilterProperties
 import ru.itmo.storage.storage.lsm.properties.LsmRepositoryFlushProperties
-import ru.itmo.storage.storage.lsm.sstable.SSTable
-import ru.itmo.storage.storage.lsm.sstable.SSTableManager
+import ru.itmo.storage.storage.lsm.core.sstable.SSTable
+import ru.itmo.storage.storage.lsm.core.sstable.SSTableManager
 import java.io.BufferedOutputStream
 import java.nio.file.Path
 import kotlin.io.path.outputStream
