--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/lsm/properties/BloomFilterProperties.kt
@@ -1,8 +1,0 @@
-package ru.itmo.storage.storage.lsm.properties
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-
-@ConfigurationProperties("storage.component.sstable.filter")
-class BloomFilterProperties {
-    var maxSize: Int = 100
-}
