--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/lsm/sstable/SSTableLoader.kt
@@ -1,5 +1,0 @@
-package ru.itmo.storage.storage.lsm.sstable
-
-interface SSTableLoader {
-    fun loadTablesSortedByCreationTimeDesc(): List<SSTable>
-}
