+++ storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/lsm/sstable/SSTableManagerImpl.kt
@@ -11,11 +11,14 @@ import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.launch
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
+import ru.itmo.storage.storage.lsm.core.AVLTree
 import ru.itmo.storage.storage.config.MEMTABLE_FLUSH
-import ru.itmo.storage.storage.lsm.MemtableService
-import ru.itmo.storage.storage.lsm.avl.AVLTree
-import ru.itmo.storage.storage.lsm.bloomfilter.BloomFilter
-import ru.itmo.storage.storage.lsm.properties.BloomFilterProperties
+import ru.itmo.storage.storage.lsm.core.MemtableService
+import ru.itmo.storage.storage.lsm.core.bloomfilter.BloomFilter
+import ru.itmo.storage.storage.lsm.core.bloomfilter.BloomFilterProperties
+import ru.itmo.storage.storage.lsm.core.sstable.SSTable
+import ru.itmo.storage.storage.lsm.core.sstable.SSTableLoader
+import ru.itmo.storage.storage.lsm.core.sstable.SSTableManager
 import ru.itmo.storage.storage.core.utils.SearchUtils
 import java.util.*
