--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/wal/entry/OperationStatus.kt
@@ -1,8 +1,0 @@
-package ru.itmo.storage.storage.wal.entry
-
-enum class OperationStatus {
-    PENDING,
-    SUCCESS,
-    FAIL,
-    UNKNOWN,
-}
