--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/wal/entry/WalEntry.kt
@@ -1,9 +1,0 @@
-package ru.itmo.storage.storage.wal.entry
-
-import java.time.Instant
-
-abstract class WalEntry(
-    val timestamp: Instant,
-    val type: WalEntryType,
-    val data: WalEntryData,
-)
