--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/wal/entry/WalEntryData.kt
@@ -1,7 +1,0 @@
-package ru.itmo.storage.storage.wal.entry
-
-import kotlinx.serialization.Serializable
-import ru.itmo.storage.storage.wal.serialization.WalEntryDataSerializer
-
-@Serializable(with = WalEntryDataSerializer::class)
-abstract class WalEntryData
