--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/wal/entry/WalEntryType.kt
@@ -1,11 +1,0 @@
-package ru.itmo.storage.storage.wal.entry
-
-enum class WalEntryType {
-    UPSERT,
-    SSTABLE_FLUSH,
-    UNKNOWN,
-}
-
-fun String.toWalEntryTypeOrUnknown(): WalEntryType = runCatching {
-    WalEntryType.valueOf(this)
-}.getOrElse { WalEntryType.UNKNOWN }
