--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/wal/entry/flush/SSTableFlushWalEntry.kt
@@ -1,10 +1,0 @@
-package ru.itmo.storage.storage.wal.entry.flush
-
-import ru.itmo.storage.storage.wal.entry.WalEntry
-import ru.itmo.storage.storage.wal.entry.WalEntryType
-import java.time.Instant
-
-class SSTableFlushWalEntry(
-    timestamp: Instant,
-    data: SSTableFlushWalEntryData,
-) : WalEntry(timestamp, WalEntryType.SSTABLE_FLUSH, data)
