--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/wal/entry/upsert/UpsertWalEntryData.kt
@@ -1,12 +1,0 @@
-package ru.itmo.storage.storage.wal.entry.upsert
-
-import kotlinx.serialization.Serializable
-import ru.itmo.storage.storage.wal.entry.OperationStatus
-import ru.itmo.storage.storage.wal.entry.WalEntryData
-
-@Serializable
-data class UpsertWalEntryData(
-    val status: OperationStatus,
-    val key: String,
-    val value: String,
-) : WalEntryData()
