--- storage-service/storage-component/src/main/kotlin/ru/itmo/storage/storage/wal/serialization/WalEntryDataSerializer.kt
@@ -1,9 +1,0 @@
-package ru.itmo.storage.storage.wal.serialization
-
-import kotlinx.serialization.json.JsonContentPolymorphicSerializer
-import kotlinx.serialization.json.JsonElement
-import ru.itmo.storage.storage.wal.entry.WalEntryData
-
-object WalEntryDataSerializer : JsonContentPolymorphicSerializer<WalEntryData>(WalEntryData::class) {
-    override fun selectDeserializer(element: JsonElement) = WalEntryData.serializer()
-}
